Hoopla updates
Filter Hoopla records according to a variety of criteria including:
- price
- age limits
- whether or not the same content exists from another provider
Fix display of hoopla records to use api rather than MARC

+++ code/java_shared_libraries/src/com/turning_leaf_technologies/dates/DateUtils.java
@@ -10,11 +10,11 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public class DateUtils {
-    public static Long getDaysSinceAddedForDate(Date curDate){
+    public static Integer getDaysSinceAddedForDate(Date curDate){
         if (curDate == null){
             return null;
         }
-        return (indexDate.getTime() - curDate.getTime()) / (1000 * 60 * 60 * 24);
+        return (int)(indexDate.getTime() - curDate.getTime()) / (1000 * 60 * 60 * 24);
     }
     private static Date indexDate = new Date();
     public static Date getIndexDate(){
